# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tayo Life Coaching API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface DimensionData {
  name: string;
  thriving: number;
  importance: number;
  openText: string;
}

export interface ChatRequest {
  messages: ChatMessage[];
  systemPrompt: string;
}

export interface ChatResponse {
  message: string;
}

export interface NarrativeRequest {
  firstName: string;
  dimensions: DimensionData[];
}

export interface NarrativeResponse {
  narrative: string;
}

export interface GeneratePlanRequest {
  conversationHistory: ChatMessage[];
  dimensions: DimensionData[];
  firstName: string;
}

export interface GeneratePlanResponse {
  plan: string;
}

export interface GenerateHabitsRequest {
  strategicPlan: string;
  firstName: string;
}

export interface GenerateHabitsResponse {
  habits: string;
}

export interface ErrorResponse {
  error: string;
}