# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tayo Life Coaching API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Send a message to the AI coach
 */
export const ChatBody = zod.object({
  messages: zod.array(
    zod.object({
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
    }),
  ),
  systemPrompt: zod.string(),
});

export const ChatResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Generate a narrative paragraph from intake data
 */
export const GenerateNarrativeBody = zod.object({
  firstName: zod.string(),
  dimensions: zod.array(
    zod.object({
      name: zod.string(),
      thriving: zod.number(),
      importance: zod.number(),
      openText: zod.string(),
    }),
  ),
});

export const GenerateNarrativeResponse = zod.object({
  narrative: zod.string(),
});

/**
 * @summary Generate the personal strategic plan
 */
export const GeneratePlanBody = zod.object({
  conversationHistory: zod.array(
    zod.object({
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
    }),
  ),
  dimensions: zod.array(
    zod.object({
      name: zod.string(),
      thriving: zod.number(),
      importance: zod.number(),
      openText: zod.string(),
    }),
  ),
  firstName: zod.string(),
});

export const GeneratePlanResponse = zod.object({
  plan: zod.string(),
});

/**
 * @summary Generate habits and goals from the strategic plan
 */
export const GenerateHabitsBody = zod.object({
  strategicPlan: zod.string(),
  firstName: zod.string(),
});

export const GenerateHabitsResponse = zod.object({
  habits: zod.string(),
});