# lib/api-zod/src/generated/types/generateHabitsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tayo Life Coaching API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateHabitsRequest {
  strategicPlan: string;
  firstName: string;
}