# lib/api-zod/src/generated/types/generatePlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tayo Life Coaching API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";
import type { DimensionData } from "./dimensionData";

export interface GeneratePlanRequest {
  conversationHistory: ChatMessage[];
  dimensions: DimensionData[];
  firstName: string;
}